Carrier configuration no longer exists
