+++ tests/test_ups.py
@@ -38,8 +38,8 @@ class TestUPS(unittest.TestCase):
         self.Account = POOL.get('account.account')
         self.Category = POOL.get('product.category')
         self.Carrier = POOL.get('carrier')
-        self.CarrierConfig = POOL.get('carrier.configuration')
         self.Party = POOL.get('party.party')
+        self.PartyConfig = POOL.get('party.configuration')
         self.PartyContact = POOL.get('party.contact_mechanism')
         self.PaymentTerm = POOL.get('account.invoice.payment_term')
         self.Country = POOL.get('country.country')
@@ -321,7 +321,7 @@ class TestUPS(unittest.TestCase):
             'currency': self.currency.id,
         }])
 
-        self.CarrierConfig.create([{
+        self.PartyConfig.create([{
             'default_validation_carrier': self.carrier.id,
         }])
